+++ src/Client.php
@@ -15,6 +15,10 @@ use SmartAssert\ServiceClient\Exception\UnauthorizedException;
 use SmartAssert\ServiceClient\Payload\UrlEncodedPayload;
 use SmartAssert\ServiceClient\Request;
 use SmartAssert\ServiceClient\Response\JsonResponse;
+use SmartAssert\WorkerClient\Factory\ApplicationStateFactory;
+use SmartAssert\WorkerClient\Factory\EventFactory;
+use SmartAssert\WorkerClient\Factory\JobCreationExceptionFactory;
+use SmartAssert\WorkerClient\Factory\JobFactory;
 use SmartAssert\WorkerClient\Model\ApplicationState;
 use SmartAssert\WorkerClient\Model\Event;
 use SmartAssert\WorkerClient\Model\Job;
