+++ src/Factory/EventFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\WorkerClient;
+namespace SmartAssert\WorkerClient\Factory;
 
 use SmartAssert\ArrayInspector\ArrayInspector;
 use SmartAssert\WorkerClient\Model\Event;
